Fix "delete" method on the service layer.
There was previously no check for the existence of a deleted object

+++ server/src/main/java/ru/ivanov/todoproject/service/ProjectService.java
@@ -80,8 +80,10 @@ public class ProjectService implements IProjectService {
     }
 
     @Override
-    public boolean deleteProject(final String userId, String projectName) throws InvalidArgumentException {
+    public boolean deleteProject(final String userId, String projectName) throws InvalidArgumentException, ObjectNotFoundException {
         if (!Validator.isArgumentsValid(userId, projectName)) throw new InvalidArgumentException();
+        final Project project = projectRepository.findProjectByName(userId, projectName);
+        if (project == null) throw new ObjectNotFoundException();
         projectRepository.deleteProject(userId, projectName);
         return true;
     }
